update frontend for development server

+++ src/pages/NotePage.js
@@ -1,18 +1,22 @@
 import React, {useState, useEffect, useContext} from "react";
+import {v4 as uuidv4} from 'uuid';
+
 import {Link} from 'react-router-dom';
 import {toast, ToastContainer} from "react-toastify";
+import {getCurrentDate} from '../utils/helpFunctions'
 
 import AuthContext from "../context/AuthContext";
+import {DivApp} from "../styles/application";
 import {ReactComponent as ArrowLeft} from '../assets/arrow-left.svg'
 import {Button, H2, H3, NoteHeader, TextArea} from "./styles/NotePageStyles";
-import {DivApp} from "../styles/application";
+
 
 const NotePage = ({match, history}) => {
     let noteId = match.params.id // shows it as a 'new'
     let [note, setNote] = useState(null) // before adding things, note will be null
-    let {authTokens} = useContext(AuthContext)
+    let {authTokens, demoUser, demoNotes} = useContext(AuthContext)
 
-    let baseURL = 'https://my-own-notes.herokuapp.com'
+    let baseURL = 'http://192.168.0.8:8000'
 
     useEffect(() => {
         getNote();
@@ -20,6 +24,15 @@ const NotePage = ({match, history}) => {
 
     let getNote = async () => {
         if (noteId === 'new') return
+        if (noteId === null) return
+
+        // Demo user has data only in cache
+        if (demoUser) {
+            let note = demoNotes.find(obj => obj.id === noteId)
+            setNote(note)
+            return
+        }
+
         let response = await fetch(`${baseURL}/api/notes/${noteId}/`, {
             method: 'GET',
             headers: {
@@ -32,6 +45,21 @@ const NotePage = ({match, history}) => {
     }
 
     let createNote = async () => {
+        if (demoUser) {
+            // create additional data for the note in cache
+            note.id = uuidv4()
+            note.updated = getCurrentDate()
+            demoNotes.push(note)
+
+            // shows the info that note has been created
+            toast.success("Note has been created!", {
+                position: toast.POSITION.TOP_RIGHT,
+                autoClose: 1500,
+            })
+            return
+        }
+
+
         let response = await fetch(`${baseURL}/api/notes/new/`, {
             method: 'POST',
             headers: {
@@ -46,7 +74,7 @@ const NotePage = ({match, history}) => {
                 position: toast.POSITION.TOP_RIGHT,
                 autoClose: 1500,
             })
-            history.push('/')
+
         } else {
             toast.error("Something gone wrong. Please try again. " +
                 "If the problem happen again, please contact support", {
@@ -55,8 +83,14 @@ const NotePage = ({match, history}) => {
         }
     }
 
-
     let updateNote = async () => {
+        if (demoUser) {
+            const oldDemoNoteIndex = demoNotes.findIndex(obj => obj.id === note.id)
+            demoNotes[oldDemoNoteIndex].body = note.body
+            demoNotes[oldDemoNoteIndex].updated = getCurrentDate()
+            return
+        }
+
         let response = await fetch(`${baseURL}/api/notes/${noteId}/`, {
             method: 'PUT',
             headers: {
@@ -65,6 +99,7 @@ const NotePage = ({match, history}) => {
             },
             body: JSON.stringify({...note})
         })
+
         await response.json()
         if (response.status === 200) {
             toast.success("Note has been saved!", {
@@ -80,6 +115,20 @@ const NotePage = ({match, history}) => {
     }
 
     let deleteNote = async () => {
+        if (demoUser) {
+            const toDeleteNoteIdx = demoNotes.findIndex(obj => obj.id === note.id)
+            // splice used to delete element from array, 1 means just one element
+            demoNotes.splice(toDeleteNoteIdx, 1)
+
+            toast.error("Note has been deleted!", {
+                position: toast.POSITION.TOP_RIGHT,
+                autoClose: 1000,
+            })
+
+            goToHomePage()
+            return
+        }
+
         await fetch(`${baseURL}/api/notes/${noteId}/`, {
                 method: 'DELETE',
                 headers: {
@@ -89,12 +138,10 @@ const NotePage = ({match, history}) => {
                 body: JSON.stringify(note)
             }
         )
-        toast.error("Note has been delated!", {
+        toast.error("Note has been deleted!", {
             position: toast.POSITION.TOP_RIGHT,
             autoClose: 1000,
         })
-        // redirect to the homepage
-        history.push('/')
     }
 
     let handleSubmit = () => {
@@ -105,7 +152,10 @@ const NotePage = ({match, history}) => {
         } else if (noteId === 'new' && note.body) {
             createNote()
         }
+        goToHomePage()
+    }
 
+    let goToHomePage = () => {
         history.push('/')
     }
 
@@ -118,19 +168,19 @@ const NotePage = ({match, history}) => {
                     </Link>
                 </H2>
                 {noteId !== 'new' ? (
-                        <H3> Edit your note </H3>
-                    ) : (
-                        <H3> Create your note </H3>
-                    )
+                    <H3> Edit your note </H3>
+                ) : (
+                    <H3> Create your note </H3>
+                )
                 }
 
                 {noteId !== 'new' ? (
                     <Button onClick={deleteNote}>Delete</Button>
-                    ) : (
-                        <Link to='/'>
-                            <Button onClick={handleSubmit}>Done</Button>
-                        </Link>
-                    )
+                ) : (
+                    <Link to='/'>
+                        <Button onClick={handleSubmit}>Done</Button>
+                    </Link>
+                )
                 }
             </NoteHeader>
